feat(refactor): VaultAccountsProvider to w3ux (#2082)
