+++ src/config/ledger.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { LedgerApp } from 'contexts/Hardware/Ledger/types';
+import type { LedgerApp } from 'contexts/LedgerHardware/types';
 import KusamaSVG from 'img/appIcons/kusama.svg?react';
 import PolkadotSVG from 'img/appIcons/polkadot.svg?react';
