--- src/contexts/Hardware/Vault/types.ts
@@ -1,20 +1,0 @@
-// Copyright 2024 @paritytech/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import type {
-  LedgerAccount,
-  VaultAccount,
-} from '@w3ux/react-connect-kit/types';
-
-export interface VaultAccountsContextInterface {
-  vaultAccountExists: (address: string) => boolean;
-  addVaultAccount: (
-    address: string,
-    index: number,
-    callback?: () => void
-  ) => LedgerAccount | null;
-  removeVaultAccount: (address: string, callback?: () => void) => void;
-  renameVaultAccount: (address: string, newName: string) => void;
-  getVaultAccount: (address: string) => LedgerAccount | null;
-  vaultAccounts: VaultAccount[];
-}
