+++ src/contexts/LedgerHardware/index.tsx
@@ -7,7 +7,7 @@ import { createContext, useContext, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { AnyJson, MaybeString } from 'types';
 import { useApi } from 'contexts/Api';
-import { getLedgerErrorType } from '../Utils';
+import { getLedgerErrorType } from './Utils';
 import { defaultFeedback, defaultLedgerHardwareContext } from './defaults';
 import type {
   FeedbackMessage,
