+++ src/library/SubmitTx/ManualSign/Ledger/Submit.tsx
@@ -6,8 +6,8 @@ import { faSquarePen } from '@fortawesome/free-solid-svg-icons';
 import type { LedgerAccount } from '@w3ux/react-connect-kit/types';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
 import { useImportedAccounts } from 'contexts/Connect/ImportedAccounts';
-import { useLedgerHardware } from 'contexts/Hardware/Ledger/LedgerHardware';
-import { getLedgerApp } from 'contexts/Hardware/Utils';
+import { useLedgerHardware } from 'contexts/LedgerHardware';
+import { getLedgerApp } from 'contexts/LedgerHardware/Utils';
 import { useNetwork } from 'contexts/Network';
 import { useTxMeta } from 'contexts/TxMeta';
 import { ButtonSubmit } from 'kits/Buttons/ButtonSubmit';
