+++ src/library/SubmitTx/ManualSign/Ledger/index.tsx
@@ -5,8 +5,8 @@ import { useEffectIgnoreInitial } from '@w3ux/hooks';
 import type { ReactNode } from 'react';
 import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useLedgerHardware } from 'contexts/Hardware/Ledger/LedgerHardware';
-import type { LedgerResponse } from 'contexts/Hardware/Ledger/types';
+import { useLedgerHardware } from 'contexts/LedgerHardware';
+import type { LedgerResponse } from 'contexts/LedgerHardware/types';
 import { useHelp } from 'contexts/Help';
 import { useTxMeta } from 'contexts/TxMeta';
 import { EstimatedTxFee } from 'library/EstimatedTxFee';
@@ -15,7 +15,7 @@ import { useImportedAccounts } from 'contexts/Connect/ImportedAccounts';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCircleExclamation } from '@fortawesome/free-solid-svg-icons';
 import { useNetwork } from 'contexts/Network';
-import { getLedgerApp } from 'contexts/Hardware/Utils';
+import { getLedgerApp } from 'contexts/LedgerHardware/Utils';
 import type { SubmitProps } from '../../types';
 import { Submit } from './Submit';
 import { ButtonHelp } from 'kits/Buttons/ButtonHelp';
