+++ src/modals/ImportLedger/Addresses.tsx
@@ -5,8 +5,8 @@ import { faArrowDown } from '@fortawesome/free-solid-svg-icons';
 import { Polkicon } from '@w3ux/react-polkicon';
 import { ellipsisFn, unescape } from '@w3ux/utils';
 import { useTranslation } from 'react-i18next';
-import { useLedgerHardware } from 'contexts/Hardware/Ledger/LedgerHardware';
-import { getLocalLedgerAddresses } from 'contexts/Hardware/Utils';
+import { useLedgerHardware } from 'contexts/LedgerHardware';
+import { getLocalLedgerAddresses } from 'contexts/LedgerHardware/Utils';
 import { usePrompt } from 'contexts/Prompt';
 import { Confirm } from 'library/Import/Confirm';
 import { Remove } from 'library/Import/Remove';
