+++ src/modals/ImportLedger/Reset.tsx
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useTranslation } from 'react-i18next';
-import { getLocalLedgerAddresses } from 'contexts/Hardware/Utils';
-import type { LedgerAddress } from 'contexts/Hardware/Ledger/types';
+import { getLocalLedgerAddresses } from 'contexts/LedgerHardware/Utils';
+import type { LedgerAddress } from 'contexts/LedgerHardware/types';
 import { usePrompt } from 'contexts/Prompt';
 import { ConfirmWrapper } from 'library/Import/Wrappers';
 import type { AnyJson } from 'types';
